Buildable

+++ glimac/include/glimac/Constants.hpp
@@ -16,7 +16,7 @@ const int NO_COLLISION_TILES[4] = {0, 1, 2, 3};
 
 
 // Timers
-const int SPEED_FACTOR = 3.5;
+const float SPEED_FACTOR = 3.5;
 
 const int SUPER_MAX   = 500 * SPEED_FACTOR;
 const int ONE_UP_MAX  = 13 * SPEED_FACTOR;
